// mods/dashboard/src/workspaces/pages/[workspace]/members/members.interfaces.ts
/**
 * Copyright (C) 2025 by Fonoster Inc (https://fonoster.com)
 * http://github.com/fonoster/fonoster
 *
 * This file is part of Fonoster
 *
 * Licensed under the MIT License (the "License");
 * you may not use this file except in compliance with
 * the License. You may obtain a copy of the License at
 *
 *    https://opensource.org/licenses/MIT
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import type { Role, WorkspaceMemberStatus } from "@fonoster/types";

export interface WorkspaceMemberDTO {
  ref: string;
  userRef: string;
  name: string;
  email: string;
  role: Role;
  status: WorkspaceMemberStatus;
  createdAt: Date;
  updatedAt: Date;
}